requirements.txt
Flask>=2.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
Pillow>=10.0.0
Werkzeug>=2.3.0
bcrypt>=4.0.0
PyYAML>=6.0
weasyprint>=60.0
opencv-python>=4.8.0